chore: Use symlinks for static assets served.
